Rename: 방 관련 entity 경로 수정
